refactor(web): split Group 2 routes (tasks/expenses/reminders/facts/links/journal/habits/recurring/watches) into APIRouters
Extract-and-recompose per the ongoing web-routes refactor: moves the main
CRUD domains out of app.py's create_app() into dedicated APIRouter modules
under ev/interfaces/web/routes/, each taking a WebContext. Also reunifies
each domain's /update endpoint (previously grouped far away, near the end of
create_app()) into the same file as its create/list/delete routes. No path,
method, or response-shape changes.
